+++ Assets/Replicator/PooledObject.cs
@@ -2,13 +2,33 @@
 
 namespace Replicator {
 	[AddComponentMenu("")] // Prevents this Component from appearing in the Unity editor.
-	internal class PooledObject : MonoBehaviour, IPooled {
+	internal class PooledObject : MonoBehaviour
+	#if UNITY_EDITOR
+	, IPooled
+	#endif
+	{
 		private ObjectPool owner;
 		private bool recycleFlag;
 
+#region HideInHierarchy
+#if UNITY_EDITOR
 		private void Awake() {
 			hideFlags = HideFlags.HideInInspector;
+			if(owner.hideUnspawned) gameObject.hideFlags |= HideFlags.HideInHierarchy;
+			UnityEditor.EditorApplication.DirtyHierarchyWindowSorting();
 		}
+
+		public void OnSpawn() {
+			if(owner.hideUnspawned) gameObject.hideFlags &= ~HideFlags.HideInHierarchy;
+			UnityEditor.EditorApplication.DirtyHierarchyWindowSorting();
+		}
+
+		public void OnRecycle() {
+			if(owner.hideUnspawned) gameObject.hideFlags |= HideFlags.HideInHierarchy;
+			UnityEditor.EditorApplication.DirtyHierarchyWindowSorting();
+		}
+#endif
+#endregion
 
 		private void LateUpdate() {
 			if(recycleFlag) {
@@ -20,14 +40,14 @@ namespace Replicator {
 		public void SetOwner(ObjectPool newOwner) {
 			if(owner == null) {
 				owner = newOwner;
-				GameObjectExtensions.poolRegistry.Add(gameObject, owner);
+				PoolRegistry.pools.Add(gameObject, owner);
 				owner.OnDisablePool += deregisterInstance;
 			}
 			else Debug.Log(Strings.SetOwnerOnOwned);
 		}
 
 		private void deregisterInstance() {
-			GameObjectExtensions.poolRegistry.Remove(gameObject);
+			PoolRegistry.pools.Remove(gameObject);
 			owner.OnDisablePool -= deregisterInstance;
 		}
 
@@ -36,9 +56,5 @@ namespace Replicator {
 		public void Recycle() => recycleFlag = true;
 
 		public GameObject Spawn(GameObject original) => owner.Spawn(transform.position, Quaternion.identity);
-
-		public void OnSpawn() => gameObject.hideFlags |= HideFlags.HideInHierarchy;
-
-		public void OnRecycle() => gameObject.hideFlags &= ~HideFlags.HideInHierarchy;
 	}
 }