+++ CHANGELOG.md
@@ -5,6 +5,25 @@ All notable changes to Replicator will be documented in this file.
 >The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Pools can be individually configured to hide/show despawned objects in the hierarchy (default: hide)
+- Particle system cleanup now allows configuration of clear behaviour
+- The behaviour to use when growing a pool is now configurable
+
+### Changed
+
+- Updated the Multi-Object pool's editor for better usability
+- Replicator is now targeting Unity 2019.3
+- Refactoring clean-ups
+- Pools' default behaviour has been changed to pre-load their full capacity. This can be disabled by editing their Pre Load option
+
+### Fixed
+
+- [#3] - Multi-pools where the underlying base variant is not available to spawn will now allow other variants to spawn
+
 ## [0.2.0-alpha]
 
 ### Added
@@ -29,3 +48,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 [0.2.0-alpha]: https://github.com/ettmetal/Replicator/compare/0.1-alpha...0.2.0-alpha
 [0.1-alpha]: https://github.com/ettmetal/Replicator/compare/a2010e58963b3f15a45031087ad54d5d1ac82bc0...0.1-alpha
+
+[#3]: https://github.com/ettmetal/Replicator/issues/3